requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
plotly==5.18.0
